requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
google-cloud-vision==3.4.5
streamlit==1.28.2
pillow==9.0.1
python-dotenv==1.0.0
langchain==0.0.333